pipeline/requirements.txt
requests==2.32.3
psycopg[binary]==3.2.13
python-dotenv==1.0.1
